move ImageQualityPercentage to main impl package as it is used in multiple media sources
